refactor(ai): centralize Gemini model config to prevent version mismatch
- Move CV parsing functions to lib/ai.ts with shared AI_CONFIG
- Update REST /api/cv/upload to use centralized parseCVWithGeminiVision/Text
- Update tRPC user.uploadCV to use same centralized functions
- Fix model version: gemini-2.0-flash → gemini-2.5-flash
- Fix ExtractedCVData import in profile page (use component type)

Single source of truth for AI model versions in AI_CONFIG.models
